requirements/prod.txt
elasticapm.contrib.django#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/prod.txt requirements.in/prod.in
#
asgiref==3.3.4
    # via django
certifi==2018.8.24
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via requests
django-audit-log-middleware==0.0.4
    # via -r requirements.in/prod.in
django-environ==0.4.5
    # via -r requirements.in/base.in
django-ipware==3.0.2
    # via
    #   django-audit-log-middleware
    #   django-log-formatter-ecs
django-log-formatter-ecs==0.0.3
    # via -r requirements.in/base.in
django-sass-processor==1.0.1
    # via -r requirements.in/base.in
django-settings-export==1.2.1
    # via -r requirements.in/base.in
django-staff-sso-client==2.2.1
    # via -r requirements.in/base.in
django-workflow-engine==0.0.14
    # via -r requirements.in/base.in
django==3.2.4
    # via
    #   -r requirements.in/base.in
    #   django-audit-log-middleware
    #   django-settings-export
    #   django-staff-sso-client
    #   django-workflow-engine
gevent==20.6.2
    # via -r requirements.in/prod.in
greenlet==0.4.16
    # via gevent
idna==2.7
    # via requests
kubi-ecs-logger==0.0.6
    # via django-log-formatter-ecs
libsass==0.20.0
    # via -r requirements.in/base.in
marshmallow==2.19.2
    # via kubi-ecs-logger
oauthlib==2.1.0
    # via requests-oauthlib
psycogreen==1.0.2
    # via -r requirements.in/prod.in
psycopg2==2.8.6
    # via -r requirements.in/base.in
pytz==2018.3
    # via django
requests-oauthlib==1.0.0
    # via django-staff-sso-client
requests==2.25.1
    # via requests-oauthlib
sentry-sdk==1.4.3
    # via -r requirements.in/base.in
six==1.15.0
    # via libsass
sqlparse==0.2.4
    # via django
urllib3==1.26.5
    # via
    #   requests
    #   sentry-sdk
whitenoise==5.2
    # via -r requirements.in/base.in
zope.event==4.4
    # via gevent
zope.interface==5.1.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
